--- a/core/src/main/scala/zio/prelude/ops/package.scala
+++ b/core/src/main/scala/zio/prelude/ops/package.scala
@@ -35,7 +35,6 @@ package object ops {
           ZPure.unit
         }
       }
-
   }
 
 }