--- a/core/src/test/scala/caliban/validation/ValidationSchemaSpec.scala
+++ b/core/src/test/scala/caliban/validation/ValidationSchemaSpec.scala
@@ -19,7 +19,7 @@ object ValidationSchemaSpec extends ZIOSpecDefault {
     gql.interpreter.exit.map(assert(_)(fails[ValidationError](hasField("msg", _.msg, equalTo(expectedMessage)))))
 
   def checkTypeError(t: __Type, expectedMessage: String): IO[ValidationError, TestResult] =
-    Validator.validateType(t).exit.map(assert(_)(fails(hasField("msg", _.msg, equalTo(expectedMessage)))))
+    Validator.validateType(t).toZIO.exit.map(assert(_)(fails(hasField("msg", _.msg, equalTo(expectedMessage)))))
 
   override def spec =
     suite("ValidationSchemaSpec")(
@@ -34,6 +34,7 @@ object ValidationSchemaSpec extends ZIOSpecDefault {
                 origin = None
               )
             )
+            .toZIO
             .exit
             .map(assert(_)(succeeds(anything)))
         },
@@ -59,6 +60,7 @@ object ValidationSchemaSpec extends ZIOSpecDefault {
                 subTypes = List(__Type(kind = __TypeKind.OBJECT))
               )
             )
+            .toZIO
             .exit
             .map(assert(_)(succeeds(anything)))
         },
@@ -297,7 +299,7 @@ object ValidationSchemaSpec extends ZIOSpecDefault {
             graphQL(resolverFieldWithArg).interpreter.exit.map(assert(_)(succeeds(anything)))
           },
           test("fields with additional nullable args are valid") {
-            Validator.validateObject(nullableExtraArgsObject).exit.map(assert(_)(succeeds(anything)))
+            Validator.validateObject(nullableExtraArgsObject).toZIO.exit.map(assert(_)(succeeds(anything)))
           },
           test("fields with additional non-nullable args are invalid") {
             checkTypeError(