Switches validation over to using ZPure from zio-prelude. The idea here being that since it is a mainly CPU bound computation we should be able to improve performance by avoiding the lifting into the ZIO environment until the last moment. Marking as draft for the moment because I want to run the benchmarks to compare against the old version.