@frekw 2023-02-27
Will be really interesting to see the benchmarks if it makes any difference!