@paulpdaniels 2023-03-04
```
Before:

[info] Benchmark                             Mode  Cnt      Score      Error  Units
[info] ValidationBenchmark.deep100          thrpt    5  36587.287 ±  597.830  ops/s
[info] ValidationBenchmark.deep1000         thrpt    5  38774.520 ±  962.654  ops/s
[info] ValidationBenchmark.deep10000        thrpt    5  37434.806 ± 1955.669  ops/s
[info] ValidationBenchmark.multifield100    thrpt    5  40530.176 ± 1331.577  ops/s
[info] ValidationBenchmark.multifield1000   thrpt    5  40654.093 ±  811.066  ops/s
[info] ValidationBenchmark.multifield10000  thrpt    5  40736.894 ±  391.088  ops/s
[info] ValidationBenchmark.simple100        thrpt    5  57284.387 ± 1203.689  ops/s
[info] ValidationBenchmark.simple1000       thrpt    5  59926.856 ±  689.663  ops/s
[info] ValidationBenchmark.simple10000      thrpt    5  60312.916 ± 2149.860  ops/s

After:

[info] Benchmark                             Mode  Cnt      Score      Error  Units
[info] ValidationBenchmark.deep100          thrpt    5  43956.629 ± 1114.713  ops/s
[info] ValidationBenchmark.deep1000         thrpt    5  41390.666 ± 1131.941  ops/s
[info] ValidationBenchmark.deep10000        thrpt    5  44897.465 ± 1168.360  ops/s
[info] ValidationBenchmark.multifield100    thrpt    5  43935.164 ± 1409.333  ops/s
[info] ValidationBenchmark.multifield1000   thrpt    5  46596.305 ± 1716.184  ops/s
[info] ValidationBenchmark.multifield10000  thrpt    5  44627.265 ± 1096.116  ops/s
[info] ValidationBenchmark.simple100        thrpt    5  62381.322 ±  957.098  ops/s
[info] ValidationBenchmark.simple1000       thrpt    5  66442.423 ± 3072.536  ops/s
[info] ValidationBenchmark.simple10000      thrpt    5  66518.309 ± 3723.042  ops/s
```