@kyri-petrou 2023-03-05
I was curious as to what difference (if any) [these changes](https://github.com/ghostdogpr/caliban/pull/1633#discussion_r1118378243) would make, and to my surprise, it's quite a lot more significant than I'd imagine. 

```
Before (as currently in this PR)
[info] GraphQLBenchmarks.fragmentsCaliban       thrpt    5     80.099 ±   0.957  ops/s
[info] execution.ValidationBenchmark.deep10000  thrpt    5  78225.327 ± 557.705  ops/s

After
[info] GraphQLBenchmarks.fragmentsCaliban       thrpt    5     98.080 ±   1.572  ops/s
[info] execution.ValidationBenchmark.deep10000  thrpt    5  88084.419 ± 484.028  ops/s
```

I only run the test suite against the fragmentsCaliban benchmark (since that query is quite beefy) and just one of the validation ones, but I guess the performance increase will be more or less the same across the other benchmarks