@kyri-petrou 2023-03-05
I rerun the 2 benchmarks above and it seems that while performance for the `ValidationXX` benchmark is roughly the same, the fragmentsBenchmark has really taken a hit (lower than current master) after switching to zio-prelude 1.0.0-RC17:

```
[info] GraphQLBenchmarks.fragmentsCaliban       thrpt    5     52.417 ±   0.273  ops/s
[info] execution.ValidationBenchmark.deep10000  thrpt    5  77426.451 ± 290.797  ops/s
```

I'm thinking it's probably due to couple of reasons:

1. The `ZPure.foreachXX` [methods](https://github.com/zio/zio-prelude/blob/series/2.x/core/shared/src/main/scala/zio/prelude/fx/ZPure.scala#L1080-L1115) use recursion instead of a while loop as in @paulpdaniels custom implementation, which is likely less performant.
2. `ZPure.when`: [the pure argument](https://github.com/zio/zio-prelude/blob/series/2.x/core/shared/src/main/scala/zio/prelude/fx/ZPure.scala#L1209) is by-value instead of by-name, which means we end up creating all the `ValidationError`s unnecessarily.

Since optimizations seem to be needed on the zio-prelude side, I'll take this discussion there. Otherwise, I'm happy with the changes in this PR - great job @paulpdaniels!