@ghostdogpr 2023-03-06
> Great job with this PR 🎉
> 
> We might want to temporarily use a custom `when` method or hold off merging this in until [zio/zio-prelude#1099](https://github.com/zio/zio-prelude/pull/1099) makes it into a new release.

+1 on this, let's have the best perf we can 😄